cluster: extract send/queue/journal path into sync_conn_write.go
Third pure code-motion increment of the #5661 sync_conn.go split. Moves
the outbound path — message queueing, the delete journal, and the send
loop plus liveness/clock-sync emitters — into a cohesive sibling file.

Moved verbatim (no rename, no logic/timing change; the moved goroutine in
sendLoop and its defers keep their exact ordering):
  PauseIncrementalSync, ResumeIncrementalSync, queueMessage,
  QueueSessionV4/V6, QueueDeleteV4/V6, armDeleteResync, journalDelete,
  flushDeleteJournal, rejournalTail, SendLivenessKeepalive, sendClockSync,
  sendLoop.

Same package (pkg/cluster); imports recomputed per file. Behavior is
byte-identical.

go build ./... passes.
